Extract document referrer script into jsplugin
