+++ app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
@@ -209,10 +209,6 @@ import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
 import org.json.JSONObject
 import timber.log.Timber
 
-private const val DUCKDUCKGO = "duckduckgo"
-private const val HTTP_SCHEME = "http"
-private const val HTTPS_SCHEME = "https"
-
 @ContributesViewModel(FragmentScope::class)
 class BrowserTabViewModel @Inject constructor(
     private val statisticsUpdater: StatisticsUpdater,
@@ -3240,26 +3236,6 @@ class BrowserTabViewModel @Inject constructor(
         val cta = currentCtaViewState().cta
         if (cta is OnboardingDaxDialogCta) {
             onDismissOnboardingDaxDialog(cta)
-        }
-    }
-
-    override fun inferLoadContext(
-        referrer: String?
-    ) {
-        Timber.d("Referrer: $referrer received in inferLoadContext")
-        val navSchemes = listOf(HTTP_SCHEME, HTTPS_SCHEME)
-        val refScheme = referrer?.toUri()?.scheme
-        val refHost = referrer?.toUri()?.host
-        if (refScheme != null && refHost != null) {
-            openerContext = when {
-                refHost.contains(DUCKDUCKGO) -> OpenerContext.SERP
-                navSchemes.any { refScheme.contains(it) } -> OpenerContext.NAVIGATION
-                refScheme.isNotEmpty() -> OpenerContext.EXTERNAL
-                else -> null
-            }
-            Timber.d("OpenerContext assigned: ${openerContext?.context} from referrer string: $referrer")
-        } else {
-            Timber.d("OpenerContext not assigned bc referrer is null ")
         }
     }
