+++ app/src/main/java/com/duckduckgo/app/browser/WebViewClientListener.kt
@@ -102,7 +102,6 @@ interface WebViewClientListener {
     fun onReceivedError(errorType: WebViewErrorResponse, url: String)
     fun recordErrorCode(error: String, url: String)
     fun recordHttpErrorCode(statusCode: Int, url: String)
-    fun inferLoadContext(referrer: String?)
 
     fun getCurrentTabId(): String
